Factor all source state into a `Sources` struct
Adds a `Sources` struct which owns the state of all of the sources,
and provides a single place to process the config and initialize all of
them.

This should also make things like `hyper` connection pools persistent
across multiple requests, which will be especially nice when we have
more than one oracle type.