Revert "terra: split oracle prevote/vote into two transactions (#71)"
This reverts commit 672a85ce8d1569f7bfa10e189d7f389c117b1de3.

Unfortunately this increases the amount of time required to broadcast
transactions which puts us outside the voting period.

We can't really fix that problem until we can query the sequence number,
as it would break local sequence number persistence.

Until we can do that, we need to look at alternatives to solving this
problem.