JSONRPC request support; MSRV 1.44+
Adds support for a JSON-serialized request body, corresponding to the
one introduced to the KMS in:

https://github.com/iqlusioninc/tmkms/pull/170

So far it just prints out the contents as a `dbg` statement, but this
can be used in integration testing to cross-check that the body is being
sent and parsed correctly.

After confirming that, we can add handling for its contents.